ported arch code
